+++ CHANGELOG.md
@@ -3,9 +3,14 @@
 Starting from Rust port of the project, all changes will be put into this file.
 
 ## Unreleased (git master)
+### New Features
+- `MU`: Support downloading with subscriptions
+
 ### Changes
 - All source: Force use `rustls` and use `http2` adaptive window for reqwest client.
-- `MU`: Update proto, move `Unrecognized` enums to the top of the list as default.
+- `MU`: Update proto information.
+- `MU`: Only use `HTTP/1.1` since `HTTP/2` broke the API request
+- `MU`: Rework downloader, image blocks are now stored to save request
 
 ### Build
 - Bump `reqwest` to 0.12